+++ spec/paperclip/content_type_detector_spec.rb
@@ -27,7 +27,13 @@ describe Paperclip::ContentTypeDetector do
   end
 
   it "returns content type of file if it is an acceptable type" do
-    allow(MIME::Types).to receive(:type_for).and_return([MIME::Type.new("application/mp4"), MIME::Type.new("video/mp4"), MIME::Type.new("audio/mp4")])
+    allow(MIME::Types).to receive(:type_for).and_return(
+      [
+        MIME::Type.new("content-type" => "application/mp4"),
+        MIME::Type.new("content-type" => "video/mp4"),
+        MIME::Type.new("content-type" => "audio/mp4")
+      ]
+    )
     allow_any_instance_of(Paperclip::ContentTypeDetector).to receive(:type_from_file_contents).and_return("video/mp4")
     @filename = "my_file.mp4"
     assert_equal "video/mp4", Paperclip::ContentTypeDetector.new(@filename).detect